sensor data comes in as millis/10 now, updater changed to reflect that.

+++ SensorData.h
@@ -11,7 +11,7 @@ struct SensorData{
 public:
 	SensorData();
 	void setID(int id){ mId = id; };
-	float nextReading();
+	int nextReading();
 private:
 	int mId;
 	int mIndex;
